@@ -4,13 +4,13 @@ import android.app.ActivityManager
 import android.app.AlertDialog
 import android.content.*
 import android.os.Build
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.provider.Settings
-import android.support.v4.content.LocalBroadcastManager
 import android.util.Log
 import android.view.MotionEvent
 import android.widget.Toast
+import androidx.localbroadcastmanager.content.LocalBroadcastManager
 import eu.sisik.overlay.OverlayService.Companion.ACTION_OVERLAY_SERVICE_STARTED
 import eu.sisik.overlay.OverlayService.Companion.ACTION_OVERLAY_SERVICE_STOPPED
 import kotlinx.android.synthetic.main.activity_main.*
@@ -64,9 +64,11 @@ class MainActivity : AppCompatActivity() {
         LocalBroadcastManager.getInstance(this).unregisterReceiver(overlayReceiver)
     }
 
-    override fun onSaveInstanceState(outState: Bundle?) {
+
+
+    override fun onSaveInstanceState(outState: Bundle) {
         super.onSaveInstanceState(outState)
-        outState?.putBoolean("isShowingExplanation", isShowingExplanation)
+        outState.putBoolean("isShowingExplanation", isShowingExplanation)
     }
 
 
